+++ src/utils/cloudinary.ts
@@ -1,5 +1,7 @@
-import imageuploader from "cloudinary";
-import fs from 'fs'
+import imageuploader from 'cloudinary';
+import fs from 'fs';
+import path from 'path';
+
 const cloudinary = imageuploader.v2;
 
 cloudinary.config({
@@ -8,32 +10,32 @@ cloudinary.config({
   api_secret: process.env.API_SECRET,
 });
 
-const uploadImageOnCloudiary = async (filePath: string | undefined , folderName: string) => {
-
-   if (!filePath) {
-      throw new Error("File path is undefined");
-    }
+const uploadImageOnCloudiary = async (filePath: string | undefined, folderName: string) => {
+  if (!filePath) {
+    throw new Error("File path is undefined");
+  }
 
   try {
     const uploadResult = await cloudinary.uploader.upload(filePath, {
       folder: folderName,
     });
-    try {
 
-      fs.unlinkSync(filePath);
-
-    } catch (error) {
-      console.log("failed to delete image from server");
+    // Ensure the file exists before attempting to delete it
+    if (fs.existsSync(filePath)) {
+      try {
+        fs.unlinkSync(filePath);
+      } catch (error) {
+        console.error("Failed to delete image from server:", error);
+      }
     }
+
     return {
-      secure_url : uploadResult.secure_url,
-      public_id :  uploadResult.public_id, 
-    }
+      secure_url: uploadResult.secure_url,
+      public_id: uploadResult.public_id,
+    };
   } catch (error: any) {
-    throw new Error(error);
+    throw new Error(`Error uploading image: ${error.message}`);
   }
-
- 
 };
 
 export { uploadImageOnCloudiary };
